refactor(cli): extract program module
